app/requirements.txt
streamlit
psycopg2-binary
pyyaml
pandas
